requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1 ; python_version >= '3.7'
aiohttp==3.9.3 ; python_version >= '3.8'
aiohttp-socks==0.8.4
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version < '3.11'
attrs==23.2.0 ; python_version >= '3.7'
certifi==2024.2.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
frozenlist==1.4.1 ; python_version >= '3.8'
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0 ; python_full_version >= '3.6.1'
hpack==4.0.0 ; python_full_version >= '3.6.1'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
idna==3.6 ; python_version >= '3.5'
jsonschema==4.21.1 ; python_version >= '3.8'
jsonschema-specifications==2023.12.1 ; python_version >= '3.8'
matrix-nio==0.24.0
multidict==6.0.5 ; python_version >= '3.7'
plemmy==0.4.1
pycryptodome==3.20.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-socks[asyncio]==2.4.4
referencing==0.33.0 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
rpds-py==0.17.1 ; python_version >= '3.8'
sqlalchemy==2.0.25
typing-extensions==4.9.0 ; python_version >= '3.8'
unpaddedbase64==2.1.0 ; python_version >= '3.6' and python_version < '4.0'
urllib3==2.2.0 ; python_version >= '3.8'
yarl==1.9.4 ; python_version >= '3.7'
